# crates/server/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Warg Registry API
  description: |
    [warg](https://warg.io/) is an open source protocol for WebAssembly component registries.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

# The Warg APIs currently do not use an authentication scheme
# Publishing requires that entries are signed with an acceptable key
...
servers:
  - url: http://localhost:8090/v1
    description: Local development server
paths:
  /fetch/ids:
    post:
      summary: Fetch package IDs
      operationId: fetchIds
      security: []
      tags:
        - fetch
      description: |
        Fetch the package IDs for registry log IDs.
      parameters:
        - name: Warg-Registry
          in: header
          description: URI Authority of another registry, if present and supported the registry will respond on behalf of another registry.
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchPackageIdsRequest"
      responses:
        "200":
          description: The package IDs were successfully fetched.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchPackageIdsResponse"
        "404":
          description: A requested entity was not found.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - status
                  - type
                  - id
                properties:
                  status:
                    type: integer
                    description: The HTTP status code for the error.
                    example: 404
                  type:
                    type: string
                    description: The type of entity that was not found.
                    enum: [log]
                    example: log
                  id:
                    type: string
                    description: The identifier of the entity that was not found.
                    example: sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c
        default:
          description: An error occurred when processing the request.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /fetch/logs:
    post:
      summary: Fetch registry logs
      operationId: fetchLogs
      security: []
      tags:
        - fetch
      description: |
        Fetch the operator and packages logs from the registry.
      parameters:
        - name: Warg-Registry
          in: header
          description: URI Authority of another registry, if present and supported the registry will respond on behalf of another registry.
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchLogsRequest"
      responses:
        "200":
          description: The logs were successfully fetched.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchLogsResponse"
        "404":
          description: A requested entity was not found.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                oneOf:
                  - "$ref": "#/components/schemas/FetchLogsIDNotFoundError"
                  - "$ref": "#/components/schemas/FetchLogsLogLengthNotFoundError"
        default:
          description: An error occurred when processing the request.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /fetch/checkpoint:
    get:
      summary: Fetch latest registry checkpoint
      operationId: getCheckpoint
      security: []
      tags:
        - fetch
      description: Fetch the latest checkpoint from the registry.
      parameters:
        - name: Warg-Registry
          in: header
          description: URI Authority of another registry, if present and supported the registry will respond on behalf of another registry.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: The checkpoint was successfully fetched.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignedCheckpoint"
        default:
          description: An error occurred when processing the request.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      description: A generic error response.
      additionalProperties: false
      required:
        - status
        - message
      properties:
        status:
          type: integer
          description: The HTTP status code for the error.
          example: 406
        message:
          type: string
          description: The error message.
          example: the server cannot produce an acceptable response
    AnyHash:
      type: string
      description: Represents a supported hash.
      example: sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c
      pattern: ^[a-z0-9-]+:[a-f0-9]+$
    FetchPackageIdsRequest:
      type: object
      description: A request to fetch package IDs from the registry.
      additionalProperties: false
      required:
        - packages
      properties:
        packages:
          type: array
          description: The log ID for each requested package.
          example: ["sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded9773"]
          minimum: 1
          items:
            $ref: "#/components/schemas/AnyHash"
            description: The log ID for each package.
    FetchPackageIdsResponse:
      type: object
      description: A response containing the requested package IDs.
      additionalProperties: false
      properties:
        packages:
          type: object
          description: The map of log ID to package ID.
          patternProperties:
            "^[a-z0-9-]+:[a-f0-9]+$":
              type: string
              nullable: true
              description: The package ID for each package. If `null`, the package ID is not able to be provided for the log ID.
          example:
            "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded9773": "example-namespace:package-name"
            "sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c": null
    FetchLogsRequest:
      type: object
      description: A request to fetch logs from the registry.
      additionalProperties: false
      required:
        - logLength
      properties:
        logLength:
          type: integer
          description: The registry checkpoint log length to fetch from.
          example: 101
          minimum: 1
        limit:
          type: integer
          description: The limit of operator and packages records to return for the fetch request.
          example: 100
          default: 100
          minimum: 1
          maximum: 1000
          format: int16
        operator:
          $ref: "#/components/schemas/AnyHash"
          description: The last known operator record fetch token.
        packages:
          type: object
          description: |
            The map of package log identifier to last known package record fetch token.

            If the last package record identifier is null, records are returned from the start of the log.
          patternProperties:
            "^[a-z0-9-]+:[a-f0-9]+$":
              $ref: "#/components/schemas/AnyHash"
              description: The last known package record identifier.
              nullable: true
          example:
            "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded9773": "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded97730"
            "sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c": null
    FetchLogsResponse:
      type: object
      description: A response containing the requested logs.
      additionalProperties: false
      properties:
        more:
          type: boolean
          description: |
            Whether there may be more records available.

            This occurs when the number of records returned for a log reaches the requested limit.

            If `true`, the client should make another request with the new last known record identifiers.
          example: false
        operator:
          type: array
          description: The operator log records for the given checkpoint since the last known record.
          maxItems: 1000
          items:
            $ref: "#/components/schemas/PublishedRecordEnvelope"
        packages:
          type: object
          description: The map of package log identifier to package records.
          patternProperties:
            "^[a-z0-9-]+:[a-f0-9]+$":
              type: array
              description: The package log records for the given checkpoint since the last known record.
              maxItems: 1000
              items:
                $ref: "#/components/schemas/PublishedRecordEnvelope"
          example:
            ? "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded97730"
            : - contentBytes: "ZXhhbXBsZQ=="
                keyId: "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded97730"
                signature: "ecdsa-p256:MEUCIQCzWZBW6ux9LecP66Y+hjmLZTP/hZVz7puzlPTXcRT2wwIgQZO7nxP0nugtw18MwHZ26ROFWcJmgCtKOguK031Y1D0="
                registryIndex: 101
                fetchToken: "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded97730"
              - contentBytes: "ZXhhbXBsZQ=="
                keyId: "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded97730"
                signature: "ecdsa-p256:MEUCIQCzWZBW6ux9LecP66Y+hjmLZTP/hZVz7puzlPTXcRT2wwIgQZO7nxP0nugtw18MwHZ26ROFWcJmgCtKOguK031Y1D0="
                registryIndex: 305
                fetchToken: "sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c"
            ? "sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c"
            : - contentBytes: "ZXhhbXBsZQ=="
                keyId: "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded97730"
                signature: "ecdsa-p256:MEUCIQCzWZBW6ux9LecP66Y+hjmLZTP/hZVz7puzlPTXcRT2wwIgQZO7nxP0nugtw18MwHZ26ROFWcJmgCtKOguK031Y1D0="
                registryIndex: 732
                fetchToken: "sha256:ygdb4e8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae00a8y"
...
    Checkpoint:
      type: object
      description: |
        A registry checkpoint.

        Checkpoints are hashed by concatenating the following:
          * A prefix of the byte string `WARG-CHECKPOINT-V0`
          * The LEB128-encoded `logLength`
          * The LEB128-length-prefixed `logRoot`
          * The LEB128-length-prefixed `mapRoot`
      additionalProperties: false
      required:
        - logLength
        - logRoot
        - mapRoot
      properties:
        logLength:
          type: integer
          description: The log length of the checkpoint.
          minimum: 1
          example: 42
        logRoot:
          $ref: "#/components/schemas/AnyHash"
          description: The log root hash of the checkpoint.
        mapRoot:
          $ref: "#/components/schemas/AnyHash"
          description: The map root hash of the checkpoint.
    TimestampedCheckpoint:
      description: |
        A timestamped registry checkpoint.

        Checkpoints are hashed by concatenating the following:
          * A prefix of the byte string `WARG-TIMESTAMPED-CHECKPOINT-V0`
          * The LEB128-encoded `logLength`
          * The LEB128-length-prefixed `logRoot`
          * The LEB128-length-prefixed `mapRoot`
          * The LEB128-encoded `timestamp`
      allOf:
        - $ref: "#/components/schemas/Checkpoint"
        - type: object
          additionalProperties: false
          required:
            - timestamp
          properties:
            timestamp:
              type: integer
              description: |
                The time that the checkpoint was generated, in seconds since
                the Unix epoch.
              minimum: 1
              example: 1692035502
    SignedCheckpoint:
      description: A signed registry checkpoint.
      allOf:
        - type: object
          required:
            - contents
          properties:
            contents:
              $ref: "#/components/schemas/TimestampedCheckpoint"
        - $ref: "#/components/schemas/Signature"
    EnvelopeBody:
      description: A signed envelope body.
      allOf:
        - type: object
          required:
            - contentBytes
          properties:
            contentBytes:
              type: string
              description: |
                Base64-encoded bytes of the content.

                The content of an envelope body is typically a serialized protocol buffer
                representing an operator or package record.
              format: byte
              maxLength: 1048576
              example: "ZXhhbXBsZQ=="
        - $ref: "#/components/schemas/Signature"
    PublishedRecordEnvelope:
      description: A signed envelope body with the published registry log index.
      allOf:
        - type: object
          required:
            - registryIndex
            - fetchToken
          properties:
            registryIndex:
              type: integer
              description: The index of the published record in the registry log.
              example: 42
            fetchToken:
              type: string
              description: The fetch token for the registry log. Used as a cursor for incrementally fetching log records.
              example: "sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c"
        - $ref: "#/components/schemas/EnvelopeBody"
        - $ref: "#/components/schemas/Signature"
    Signature:
      type: object
      description: Represents a signature of content.
      required:
        - keyId
        - signature
      properties:
        keyId:
          $ref: "#/components/schemas/AnyHash"
          description: The signing key identifier.
        signature:
          type: string
          description: The algorithm-prefixed bytes of the signature (base64 encoded).
          pattern: ^[a-z0-9-]+:(?:[A-Za-z0-9+\/]{4})*(?:[A-Za-z0-9+\/]{4}|[A-Za-z0-9+\/]{3}=|[A-Za-z0-9+\/]{2}={2})$
          example: "ecdsa-p256:MEUCIQCzWZBW6ux9LecP66Y+hjmLZTP/hZVz7puzlPTXcRT2wwIgQZO7nxP0nugtw18MwHZ26ROFWcJmgCtKOguK031Y1D0="
...
    FetchLogsIDNotFoundError:
        type: object
        additionalProperties: false
        required:
          - status
          - type
          - id
        properties:
          status:
            type: integer
            description: The HTTP status code for the error.
            example: 404
          type:
            type: string
            description: The type of entity that was not found.
            enum: [log, fetchToken]
            example: log
          id:
            type: string
            description: The identifier of the entity that was not found.
            example: sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c
    FetchLogsLogLengthNotFoundError:
        type: object
        additionalProperties: false
        required:
          - status
          - type
          - id
        properties:
          status:
            type: integer
            description: The HTTP status code for the error.
            example: 404
          type:
            type: string
            description: The type of entity that was not found.
            enum: [logLength]
            example: log
          id:
            type: integer
            description: The log length that was not found.
            example: 1001
...